+++ src/mcmu/__main__.py
@@ -8,33 +8,29 @@ from pathlib import Path  # Import for file functions
 from logging import getLogger, basicConfig  # Logging functionality
 from requests import get  # Get files from the CDN
 from argparse import ArgumentParser  # Command line arguments class
-from .ModrinthAPI import ModrinthAPI  # Modrinth API code (Local import)
-
-__version__ = "2.0.0.dev3"
-__author__ = "Josiah Jarvis"
+from importlib.metadata import version as get_version
 
 logger = getLogger(__name__)
 basicConfig(format="%(levelname)s: %(message)s")  # Set logging format
 
 parser = ArgumentParser(
     prog="mcmu",
-    epilog=f"Version: {__version__}",
-    description="Downloads Minecraft mods from Modrinth"
+    description="Downloads Minecraft mods from Modrinth",
+    epilog=f"Version: {get_version('mcmu')}"
 )
-group = parser.add_mutually_exclusive_group()  # Get mutually exclusive group set up
+group = parser.add_mutually_exclusive_group()  # Group for arguments
 group.add_argument("-u", "--update", help="Updates installed mods", action="store_true")
 group.add_argument("-r", "--remove", help="Removes an installed mod")
 group.add_argument("-i", "--install", help="Installs a mod")
 group.add_argument("-l", "--list", help="List installed mods", action="store_true")
 group.add_argument("-s", "--search", help="Search packages on Modrinth")
+group.add_argument("-d", "--dependency", help="List a mods dependency's")
 # Unix systems defaults to ~/.minecraft/ for the minecraft dir, I don't know about Windows or MacOS
 parser.add_argument("-m", "--minecraft_dir", default=Path(Path.home(), ".minecraft/"), help="Path to the Minecraft folder, defaults to '~/.minecraft/'")
 # Game version defaults to 1.21.11 as it is the latest Minecraft release
 parser.add_argument("-g", "--game_version", default="26.1", help="The game version to use to install mods, defaults to '26.1'")
 
 args = parser.parse_args()  # Parser the arguments
-
-ModAPI = ModrinthAPI(f"Josiah-Jarvis/MCMU/{__version__} (https://github.com/Josiah-Jarvis/MCMU)")
 
 
 def check_update(mod_name: str, current_version: str) -> [bool, dict]:
@@ -69,8 +65,7 @@ def check_update(mod_name: str, current_version: str) -> [bool, dict]:
 def list_mods(mod_path: Path):
     mods = {}
     for mod in listdir(mod_path):
-        pattern = r'^(.*?)_version_(.*)\.jar$'
-        m = match(pattern, mod)
+        m = match(r'^(.*?)_version_(.*)\.jar$', mod)
         mods[m.group(1)] = {
             "name": m.group(1),
             "version": m.group(2),
@@ -91,6 +86,106 @@ def install_mod(file: str, path: Path):
     return True
 
 
+class ModrinthAPI:
+    def __init__(self, UserAgent: str):
+        self.headers = {'User-Agent': UserAgent}
+
+    def search(self, query: str, facets: str) -> [dict, int]:
+        """Search's Mod on Modrinth: https://docs.modrinth.com/api/operations/searchprojects/
+
+        Arguments:
+            query -- The query string
+            facets -- Used to limit the search results see Modrinth API documentation
+
+        Returns:
+            dict: Response data
+            int:
+                410: API deprecated
+                400: Search invalid
+        """
+        parameters = {
+            'query': query,
+            'facets': facets,
+            'limit': "100"
+        }
+        response = get(url="https://api.modrinth.com/v2/search", params=parameters, headers=self.headers)
+        if response.status_code != 200:
+            return response.status_code
+        else:
+            return response.json()
+
+    def project(self, slug: str) -> [dict, int]:
+        """Get data about a project: https://docs.modrinth.com/api/operations/getproject/
+
+        Arguments:
+            slug -- The slug of the project
+
+        Returns:
+            dict: Project data
+            int:
+                410: API deprecated
+                404: Project not found
+        """
+        response = get(url=f"https://api.modrinth.com/v2/project/{slug}", headers=self.headers)
+        if response.status_code != 200:
+            return response.status_code
+        else:
+            return response.json()
+
+    def project_dependencies(self, slug: str) -> [dict, int]:
+        """Get a list of a projects dependencies: https://docs.modrinth.com/api/operations/getdependencies/
+
+        Arguments:
+            slug -- The slug of the project
+
+        Returns:
+            dict: Dependency information
+            int:
+                410: API deprecated
+                404: Project not found
+        """
+        response = get(url=f"https://api.modrinth.com/v2/project/{slug}/dependencies", headers=self.headers)
+        if response.status_code != 200:
+            return response.status_code
+        else:
+            return response.json()
+
+    def project_version(self, slug: str, loaders: str, game_version: str) -> [dict, int]:
+        """List a projects versions: https://docs.modrinth.com/api/operations/getprojectversions/
+
+        Arguments:
+            slug -- The slug of the project
+            loaders -- Loaders to filter for
+            game_version -- Game versions to filter for
+
+        Returns:
+            dict: Project version data
+            int:
+                410: API deprecated
+                404: Project not found
+        """
+        parameters = {
+            'loaders': loaders,
+            'game_versions': game_version,
+            'include_changelog': 'false'
+        }
+        response = get(url=f"https://api.modrinth.com/v2/project/{slug}/version", params=parameters, headers=self.headers)
+        if response.status_code != 200:
+            return response.status_code
+        else:
+            return response.json()
+    
+    def get_project_version(self, slug: str, version_number: str, loaders: str, game_version: str) -> [dict, int]:
+        response = get(url=f"https://api.modrinth.com/v2/project/{slug}/version/{version_number}", params=parameters, headers=self.headers)
+        if response.status_code != 200:
+            return response.status_code
+        else:
+            return response.json()  
+
+
+ModAPI = ModrinthAPI(f"Josiah-Jarvis/MCMU/{get_version('mcmu')} (https://github.com/Josiah-Jarvis/MCMU)")
+
+
 def main():
     """Main function
 
@@ -106,9 +201,7 @@ def main():
     if args.update:
         for mod_name in mods:
             latest_version = check_update(mod_name, mods[mod_name]['version'])  # Check for update
-            if latest_version == 404:  # Thats weird
-                print(f"Mod: {mod_name}does not exist on Modrinth.")
-            elif latest_version is None:  # If latest version is None no newer version was found
+            if latest_version is None:  # If latest version is None no newer version was found
                 print("No version found for the specified game version and loader.")
             elif latest_version:  # If latest version is a dict it should return True
                 old_file = Path(mod_path, mods[mod_name]['file'])  # Path to the old mod file
@@ -116,7 +209,7 @@ def main():
                 if input(f"{mods[mod_name]['name']} will take up: {additional_storage} additional bytes, would you like to install? [Y/n]: ") is ("" or "Y"):  # Ask them if they want to install it
                     for dependency in latest_version['dependencies']:
                         mod_data = ModAPI.project(dependency['project_id'])
-                        if (not mod_data['slug'] in mods) and (dependency['dependency_type'] == "required"):
+                        if (mod_data['slug'] not in mods) and (dependency['dependency_type'] == "required"):
                             dependency_latest_version = check_update(mod_data['slug'], 0)
                             mod_jar_file = Path(mod_path, f"{mod_data['slug']}_version_{dependency_latest_version['version_number']}.jar")
                             if install_mod(dependency_latest_version['files'][0]['url'], mod_jar_file):
@@ -124,7 +217,7 @@ def main():
                             else:
                                 print("Failed to download required dependency.")
                                 return 1
-                        elif (not mod_data['slug'] in mods) and (dependency['dependency_type'] == "optional"):
+                        elif (mod_data['slug'] not in mods) and (dependency['dependency_type'] == "optional"):
                             dependency_latest_version = check_update(mod_data['slug'], 0)
                             print(f"Optional dependency: {mod_data['slug']} not installed")
                         elif (mod_data['slug'] in mods) and (dependency['dependency_type'] == "incompatible"):
@@ -155,7 +248,7 @@ def main():
             if input(f"{args.install} will take up: {latest_version['files'][0]['size']} bytes, would you like to install? [Y/n]: ") is ("" or "Y"):  # Ask if the want to install it
                 for dependency in latest_version['dependencies']:
                     mod_data = ModAPI.project(dependency['project_id'])
-                    if (not mod_data['slug'] in mods) and (dependency['dependency_type'] == "required"):
+                    if (mod_data['slug'] not in mods) and (dependency['dependency_type'] == "required"):
                         dependency_latest_version = check_update(mod_data['slug'], 0)
                         mod_jar_file = Path(mod_path, f"{mod_data['slug']}_version_{dependency_latest_version['version_number']}.jar")
                         if install_mod(dependency_latest_version['files'][0]['url'], mod_jar_file):
@@ -163,7 +256,7 @@ def main():
                         else:
                             print("Failed to download required dependency.")
                             return 1
-                    elif (not mod_data['slug'] in mods) and (dependency['dependency_type'] == "optional"):
+                    elif (mod_data['slug'] not in mods) and (dependency['dependency_type'] == "optional"):
                         dependency_latest_version = check_update(mod_data['slug'], 0)
                         print(f"Optional dependency: {mod_data['slug']} not installed")
                     elif (mod_data['slug'] in mods) and (dependency['dependency_type'] == "incompatible"):
@@ -209,6 +302,15 @@ def main():
         else:
             for mod in response['hits']:  # Iterate over and list all the mods
                 print(f"{mod['title']}:\n\tDescription: {mod['description']}\n\tAuthor: {mod['author']}\n\tDownloads: {mod['downloads']}\n\tLatest Version: {mod['latest_version']}\n\n")
+    elif args.dependency:
+        response = ModAPI.project_dependencies(args.dependency)
+        if response == 410:  # That means the API is deprecated: https://docs.modrinth.com/api/
+            logger.critical("API is deprecated, try updating MCMU")
+        elif response == 404:  # Response of 400 means request was invalid: https://docs.modrinth.com/api/operations/searchprojects/
+            logger.error("Mod not found on Modrinth")  # Print error
+        else:
+            for mod in response['projects']:
+                print(f"{mod['title']}:\n\tDescription: {mod['description']}\n\tDownloads: {mod['downloads']}\n\n")
     else:
         parser.print_help()  # Prints help message if there is nothing to do